pyproject.toml
[project]
name = "nonebot_plugin_uptime_kuma_puller"
version = "0.1.0"
description = "This is a plugin that can generate a UptimeKuma status page summary for your Nonebot"
authors = [
  { name = "HTony03", email = "HTony03@foxmail.com" },
  { name = "bananaxiao2333", email = "banana.xiao@qq.com" },
]
readme = "README.md"
keywords = ["nonebot", "kuma", "qqbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
license = "MIT"

dependencies = [
    "aiohttp~=3.11",
    "nonebot2~=2.3",
    "nonebot-plugin-waiter~=0.8.1",
]

[project.urls]
homepage = "https://github.com/bananaxiao2333/nonebot-plugin-uptime-kuma-puller"
repository = "https://github.com/bananaxiao2333/nonebot-plugin-uptime-kuma-puller"
changelog = "https://github.com/bananaxiao2333/nonebot-plugin-uptime-kuma-puller/releases"
issue = "https://github.com/bananaxiao2333/nonebot-plugin-uptime-kuma-puller/issues"
